+++ apps/elf-eval/src/bin/real_world_job_benchmark/quantitative/report.rs
@@ -1,22 +1,13 @@
 mod controls;
+mod input;
 mod row;
 
+pub(crate) use self::input::QuantitativeReportInput;
+
 use crate::{
-	AdapterReport, JobReport, Path, QuantitativeBenchmarkReport, RealWorldJob, ReportSummary,
-	Result,
+	QuantitativeBenchmarkReport, Result,
 	quantitative::{self, QUANTITATIVE_K_VALUES, QUANTITATIVE_SCOREBOARD_SCHEMA, product_manifest},
 };
-
-pub(crate) struct QuantitativeReportInput<'a> {
-	pub(crate) run_id: &'a str,
-	pub(crate) generated_at: &'a str,
-	pub(crate) adapter: &'a AdapterReport,
-	pub(crate) source_jobs: &'a [RealWorldJob],
-	pub(crate) jobs: &'a [JobReport],
-	pub(crate) summary: &'a ReportSummary,
-	pub(crate) product_manifest_path: Option<&'a Path>,
-	pub(crate) audit_manifest_path: Option<&'a Path>,
-}
 
 pub(crate) fn quantitative_scoreboard_report(
 	input: QuantitativeReportInput<'_>,
